ST6RI-161 Initial derived state computing for generalizations
This change adds a derived state computer to the Xtext resources that
generates Generalization (including Subsetting, FeatureTyping and
Superclassing) instances after parsing finished.

The current status is incomplete as the following issues are to be
fixed:

 * Dependency relations needs to be managed explicitly, e.g. feature
types depend on the types of subsetted feature types or the original
types of conjugator features. To handle these cases properly, the
traversal of the types has to be ordered by the dependency relations
instead of the current unspecified order.
 * It is not possible to distinguish between implied and explicitly
declared generalizations that is required for some calculations. Some
cases can be handled by appropriate ordering: calculate the results
_before_ the derived relations are added but this might not be always
possible because of the dependency ordering. One such case might be the
case of Feature::types reference.