ST6RI-204 Ensures demonstrator projects do not need to be Java projects
Xtext by default adds Java project specific container management options
into the UI modules of projects; this change replaces that container
manager with an Eclipse project specific one and updates the bundled
example projects not to be a Java-project anymore.

Furthermore, a new section was added to the readme file that describes
the steps to initialize a new project that can reference model elements
from other projects.