ST6RI-193 Removes some unnecessary cleanup during linking
Each time the linker implementation tries to resolve references in a
resource it starts by clearing existing proxies. However, in some cases
it gets confused by the SysML metamodel implementations that are derived
but not marked as such, e.g. in case of Relationship#source and
Relationship#target. Furthermore, removing these references can result
in reentrant resolution that confuses Xtext even more.

This change updates the linker implementation of Xtext to not to clear
these features, as it is unnecessary (the superclass implementation is
not used for now, while the subclasses always build new lists each time
they are called).