ST6RI-212 Fixes a regression related to resolving operator '..'
When switching resolution to rely more on the Xtext index, a regression
was introduced when resolving the operator '..' making the operator not
resolve from the standard library. More precisely,
OperatorExpressionImpl replaces the names '.' and '..' to match earlier
behavior of the name resolution. However, after ST6RI-206 was fixed it
is possible to simply remove this workaround and query the index using
the original name.