ST6RI-212 Adds evaluated files to index
This change adds new resources loaded by SysMLInteractive::parse to the
index. The previous versions missing these call resulted in references
not resolved correctly between files added via the 'eval' functionality
of the interactive parser.

Furthermore, the launch configuration for the profiler was extended to
include a test case like this as well.