ST6RI-233 Re-implements name shadowing in interactive parser
When updating the interactive parser to fill Xtext index for the used
resource set (to speed up lookup), the name shadowing logic expected by
the interactive parser broke as the default Xtext index lookup does not
expects functional differences between elements of the same qualified
name returned from separate files so it always has returned one of the
required elements.

This changes customizes the resource description data implementation
used in SysMLUtil by adding an explicitly ordered set for lookup that
knows about the order the resources need to be considered.

NOTE that the current ordering in not appropriate to be merged as (1) it
uses string compare instead of numeric compare, and (2) the uri-based
ordering might not be appropriate for other (non-Jupyter) standalone
environments.