ST6RI-210 Avoid excessive logging of unnecessary messages
The removal of implicit generalizations from the resources caused Xtext
index to throw messages that the Feature::typing references to these
generalizations are not pointing to anything stored in the EMF
containment tree but otherwise ignore the problem. This commit updates
the indexing strategy of KerML and SysML languages to avoid touching the
problematic references thus it avoids logging the error message.