ST6RI-282 Replaces most usages of EObjectEList with UnmodifiableEList
A side effect of these changes is that collections returned cannot be
modified anymore as an exception is thrown. However, earlier modifying
these lists (1) sent notifications, possibly resulting in unexpected
results for listeners, and (2) these modifications were not synchronized
with the underlying models - in other words, modifying these lists
earlier is usually problematic. Modifications inside the EMF model (e.g.
when updating the calculation in subtypes) were updated not to rely on
updating the parent list (where necessary).