ST6RI-210 Executes transformation steps from a derived state computer
This change register a derived state computer (together with the
required Resource, IResourceValidator and resource description strategy
manager that calls the transform method on each element (after indexing
is complete).

Furthermore, the test case IndividualUsage_Invalid needed to be updated
to make sure the required Performances::literalEvaluations element is
available for type resolution.