ST6RI-316 Makes TypeFeaturing instances implicit
This commit updates the handling of the TypeFeaturing instances created
by Feature::addFeaturingType implicit, similar to the solution
introduced for ST6RI-210 for generalizations. Given this feature is not
a generalization, it is stored in a separate collection (not as part of
the implicit general type map of Element (although it could have been
reused).

The commit also adds the generation to the transformation in ElementUtil
to make the XMI serialization the same as before (with the exception
that now implicit generalizations are consistently added before these
TypeFeaturing instances).