ST6RI-327 Adds implicit binding connectors to the outline
Given implicit binding connectors are represented with multiple elements
in the tree, some of them are also implicitly generated, adding these
items required a few non-trivial changes in the outline:

 1. The default traversal logic was updated: for implicit elements (that
are not stored in the containment hierarchy) all calls are redirected
through a newly introduced helper method for creating new tree items
that select the default EObjectNode and the ImplicitNodes accordingly.
 2. For implicit nodes it also  calls the transform method to make sure
every derived implicit element is available.