ST6RI-318 Adds xpect test execution to the build
This change executes the Xpect tests for the KerML and SysML languages
from the Maven build. The tests don't run correctly from PDE
environments, so instead of the tycho-test-runner the default surefire
runner was configured for the tests.

A test failure will mark the build as failed and the build log would
contain some information about the cause. The detailed test reports are
available from the target/surefire reports of both test projects
(including execution times, etc.).

Finally, to avoid potential issues with the availability of the Xpect
repository, the tests were added to a dedicated Maven profile, making it
possible to skip the test execution by specifying the property
`xpect.tests.skip`. E.g. running `mvn package -Dxpect.tests.skip` builds
the corresponding package, but does not try to download Xpect and run
the tests.