ST6RI-305 Changed formatRootElement to only include owned members.
This means that the output of processing a cell in Jupyter will now only
list the new elements defined in that cell, not elements that are
imported at the top level. This makes sense because top-level imported
elements are not currently re-exported and can't be referenced from
other cells (due to the way global scoping is implemented).