ST6RI-823 Corrects fix for incorrect removal of templating-maven-plugin
The templating-maven-plugin, used to copy the resources into the target
directory `kernel`, was causing the kernel zips recursively containing
earlier builds as well, causing its size to balloon (mostly because
maven considered a target folder as a source folder). However, simply
removing broke clean builds, as the installer and other metadata were
missing from the output folder.

This commit uses the maven-resources-plugin to copy the required files
to the installer, thus avoiding the original issue.