SYSML21-322 Fixed diamond inheritance problems in Systems Library.
Mostly, this involved just adding appropriate explicit redefinitions.
But:

- Flows was fixed in part by updating Connections::BinaryConnection and
Transfers::FlowTransferBefore to re-order specializations.

- Usages typed by metadata definitions in SysML were mostly fixed by the
fix to Metadata. However, this did not fix usages typed by KerML
metaobjects. To consistently resolve all problems, "subsets
Metadata::metadataItems" was added to all metadata/metaobject item
usages.

- Views was fixed by adding a redefinition of "this" to
View::viewpointConformance::viewpointSatisfactions, but with a default
of "that.that" (which is basically the default value of "this" for
View::viewpointSatisfactions; but see also SYSML21-323).